requirements.txt
# Core Requirements
fpdf2==2.7.4  # Lightweight PDF generation (fork of FPDF)

# Development/Extras (uncomment if needed)
# pytest==7.4.0  # For running unit tests
# python-dotenv==1.0.0  # For environment variables
# docker==6.1.3  # For Docker integration